fk u paul co suck my dick

+++ src/main/Attackable.java
@@ -3,5 +3,5 @@ package main;
 public interface Attackable
 {
     void attack(Attackable attackable);
-    void damage(int damage);
+    void damage(float damage);
 }
